+++ mensasverige/app/(tabs)/_layout.tsx
@@ -4,6 +4,7 @@ import { Platform } from 'react-native';
 
 import { HapticTab } from '@/components/HapticTab';
 import { IconSymbol } from '@/components/ui/IconSymbol';
+import { MaterialIcons } from '@expo/vector-icons';
 import TabBarBackground from '@/components/ui/TabBarBackground';
 import { Colors } from '@/constants/Colors';
 import { useColorScheme } from '@/hooks/useColorScheme';
@@ -29,15 +30,29 @@ export default function TabLayout() {
       <Tabs.Screen
         name="(home)"
         options={{
-          title: 'Home',
+          title: 'Information',
           tabBarIcon: ({ color }) => <IconSymbol size={28} name="house.fill" color={color} />,
         }}
       />
       <Tabs.Screen
-        name="explore"
+        name="schedule"
         options={{
-          title: 'Explore',
-          tabBarIcon: ({ color }) => <IconSymbol size={28} name="paperplane.fill" color={color} />,
+          title: 'Schema',
+          tabBarIcon: ({ color }) => <MaterialIcons name="schedule" size={28} color={color} />,
+        }}
+      />
+      <Tabs.Screen
+        name="(map)"
+        options={{
+          title: 'Karta',
+          tabBarIcon: ({ color }) => <MaterialIcons name="map" size={28} color={color} />,
+        }}
+      />
+      <Tabs.Screen
+        name="event"
+        options={{
+          title: 'Spontant',
+          tabBarIcon: ({ color }) => <MaterialIcons name="celebration" size={28} color={color} />,
         }}
       />
     </Tabs>
