+++ mensasverige/app/(tabs)/event.tsx
@@ -1,0 +1,23 @@
+import { StyleSheet } from 'react-native';
+
+import ParallaxScrollView from '@/components/ParallaxScrollView';
+import { ThemedText } from '@/components/ThemedText';
+import { ThemedView } from '@/components/ThemedView';
+
+export default function EventScreen() {
+  return (
+    <ParallaxScrollView>
+      <ThemedView style={styles.titleContainer}>
+        <ThemedText type="title">Event</ThemedText>
+      </ThemedView>
+    </ParallaxScrollView>
+  );
+}
+
+const styles = StyleSheet.create({
+  titleContainer: {
+    flexDirection: 'row',
+    alignItems: 'center',
+    gap: 8,
+  },
+});
